+++ ichnaea/webapp/config.py
@@ -24,6 +24,7 @@ from ichnaea.log import (
     configure_raven,
     configure_stats,
 )
+from ichnaea.webapp import renderers
 from ichnaea.webapp.monitor import configure_monitor
 
 
@@ -111,6 +112,9 @@ def main(app_config, ping_connections=False,
 
     # Add special JSON renderer with nicer float representation
     config.add_renderer('floatjson', floatjson.FloatJSONRenderer())
+
+    # Add text-as-JS renderer.
+    config.add_renderer('js', renderers.JSRenderer())
 
     # freeze skip logging set
     config.registry.skip_logging = frozenset(config.registry.skip_logging)